Move subscriber to it's own file
