+++ pubsubs/__init__.py
@@ -1,4 +1,3 @@
 from pubsubs.mq import MessageQueue
-from pubsubs.mq import Subscriber
 
 from . import backends
