+++ pubsubs/mq.py
@@ -3,19 +3,24 @@ from abc import abstractmethod
 
 from interface_meta import InterfaceMeta
 
+from pubsubs.subscriber import Subscriber
+
 
 class MessageQueue(metaclass=InterfaceMeta):
 
     BACKENDS = None
 
-    def __init__(self, name, registry, backend, listeners, **kwargs):
+    def __init__(self, name, registry, backend, listeners, **config):
+        #: name registered under
         self._name = name
+        #: implementation config
+        self.config = config
+        #: name of implementation
         self.backend = backend
+        #: the registry associated with this implementation
         self.registry = registry
+        #: servers to bind to
         self.listeners = listeners
-
-        self.config = kwargs
-        self._subscriber = self._assign_subscriber()
 
     @classmethod
     def __register_implementation__(cls):
@@ -46,13 +51,10 @@ class MessageQueue(metaclass=InterfaceMeta):
         self._connect()
         topics = list(topics)
 
-        # Create a new subscriber from message queue implementation
-        return self._subscriber(
+        # Create a new subscriber from the message queue implementation
+        return Subscriber.for_backend(self.backend)(
             self._config, topics, self.serializer, **self._sub_config
-        )
-
-    def _assign_subscriber(self):
-        return Subscriber.for_backend(self.backend)
+        ).connect()
 
     def publish(self, topic, message):
         self._connect()
@@ -65,50 +67,4 @@ class MessageQueue(metaclass=InterfaceMeta):
 
     @abstractmethod
     def _connect(self):
-        raise NotImplementedError
-
-
-class Subscriber(metaclass=InterfaceMeta):
-
-    BACKENDS = None
-
-    def __init__(self, config, topics, serializer, **kwargs):
-        self.config = config
-        self.topics = topics
-        self.serializer = serializer
-        self.subscriber_config = kwargs
-        self._connect()
-
-    @abstractmethod
-    def _connect(self):
-        raise NotImplementedError
-
-    @classmethod
-    def __register_implementation__(cls):
-        if not hasattr(cls, "_backends"):
-            cls._backends = {}
-
-        cls._backends[cls.__name__] = cls
-
-        registry_keys = getattr(cls, "BACKENDS", []) or []
-        if registry_keys:
-            for key in registry_keys:
-                if key in cls._backends and cls.__name__ != cls._backends[key].__name__:
-                    raise NameError("Key already registered")
-                else:
-                    cls._backends[key] = cls
-
-    @classmethod
-    def for_backend(cls, backend):
-        if backend not in cls._backends:
-            raise KeyError(f"Missing '{backend}' implementation")
-        return functools.partial(cls._backends[backend], backend=backend)
-
-    def listen(self):
-        """ Blocks until a message is received."""
-        message = next(self)
-        return self.serializer(message)
-
-    @abstractmethod
-    def __next__(self):
         raise NotImplementedError
